fix!(pdf): Corregir el tipo de argumento en la llamada a setTextMatrix
Se ha corregido un error de compilación (`Type mismatch`) en la función `drawWrappedText`.

La causa era que se estaban pasando literales de tipo `Float` (ej. `1f`) y variables `Float` a `contentStream.setTextMatrix`, mientras que la API de esta versión de PDFBox requiere argumentos de tipo `Double`.

Se han convertido todos los argumentos a `Double` para que coincidan con la firma del método, solucionando el error de compilación.

+++ app/src/main/java/com/example/dynamiccollage/utils/PdfGenerator.kt
@@ -128,7 +128,8 @@ object PdfGenerator {
                     else -> bounds.lowerLeftX
                 }
                 // Using setTextMatrix for absolute positioning, which is safer.
-                contentStream.setTextMatrix(1f, 0f, 0f, 1f, startX, y)
+                // The API expects Doubles for the matrix values.
+                contentStream.setTextMatrix(1.0, 0.0, 0.0, 1.0, startX.toDouble(), y.toDouble())
                 contentStream.showText(line)
                 contentStream.endText()
                 y -= leading
